refactor: simplify filenames
