+++ src/storage/sequelize.ts
@@ -1,6 +1,12 @@
 import { UmzugStorage } from './contract';
-import { SequelizeType, ModelClassType } from './type-helpers/sequelize-type-helpers';
 import { SetRequired } from 'type-fest';
+import { Sequelize as SequelizeType, Model as ModelClass } from 'sequelize';
+
+interface ModelTempInterface extends ModelClass {
+	[key: string]: any;
+}
+
+type ModelClassType = typeof ModelClass & (new (values?: object, options?: any) => ModelTempInterface);
 
 // eslint-disable-next-line @typescript-eslint/class-name-casing
 interface _SequelizeStorageConstructorOptions {
